move misc.io to numerics.io
