+++ pyphs/dictionary/pwl/_dissipative.py
@@ -12,7 +12,7 @@ import numpy as np
 import sympy as sp
 from .. import edges
 from pyphs import Graph
-from pyphs.misc.io import data_generator
+from pyphs.numerics.io import data_generator
 from ..tools import symbols
 from .tools import pwl_func
