+++ pyphs/dictionary/pwl/_storage.py
@@ -11,7 +11,7 @@ from __future__ import absolute_import, division, print_function
 import numpy as np
 from .. import edges
 from pyphs import Graph
-from pyphs.misc.io import data_generator
+from pyphs.numerics.io import data_generator
 from ..tools import symbols
 from .tools import pwl_func
